test-requirements.txt
# needed for doc build
docutils>=0.11,!=0.13.1  # OSI-Approved Open Source, Public Domain
openstackdocstheme>=2.0.0  # Apache 2.0
sphinx!=2.1.0,>=2.0
PyYAML>=3.1.0
yamllint>=0.5.2
requests-cache
Pillow>=2.4.0 # PIL License
hacking>=3.0,<3.1.0 # Apache-2.0
whereto>=0.3.0  # Apache-2.0

fixtures>=3.0.0 # Apache-2.0/BSD
python-subunit>=0.0.18 # Apache-2.0/BSD
stestr>=2.0.0 # Apache-2.0
testscenarios>=0.4 # Apache-2.0/BSD
testtools>=1.4.0 # MIT
oslotest>=1.10.0 # Apache-2.0

coverage>=4.0 # Apache-2.0
